cloud_server/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
# 不固定 pymongo 具体小版本，避免镜像源缺少 4.6.1 导致构建失败
pymongo
paho-mqtt==1.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.4
Pillow==10.3.0